+++ src/app/__tests__/Button.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { Button } from '../src/shared/ui/Button/index';
+import { Button } from '../../shared/ui/Button/index';
 import React from 'react';
-import { eButtonType } from '../src/shared/utils/data';
+import { eButtonType } from '../../shared/utils/data';
 import { test, expect } from 'vitest';
 
 test('renders button with correct text', () => {
