+++ src/app/__tests__/Input.test.tsx
@@ -1,7 +1,7 @@
 import { render, fireEvent } from '@testing-library/react';
 import { test, expect } from 'vitest';
 import React from 'react';
-import Input from '../src/shared/ui/Input/index';
+import Input from '../../shared/ui/Input/index';
 
 test('renders Input component with label, placeholder, and handles input', () => {
   const { getByPlaceholderText } = render(
